requirements.txt
pytesseract
pyserial
numpy
matplotlib
pillow
opencv-python
jamspell
